refactor: unify database bootstrap
